Parse casts
These are tricky because of ambiguity. For example “3 * (m) * 2” and “3
* (m) 2”. In the first case m is a variable and both operations are
multiplies. In the second case, m is a type used to cast 2. The
ambiguity lies in determining if (m) is an access or a cast operator.
To solve this, we attempt to parse a unary after parsing a possible
cast operator. If this fails then the operator is actually an access.