Add ParserException and ParseFailure
Parse failures happen when there is nothing to parse that matches the
syntax. A parse error happens when the syntax was not valid for the
match. For example, when “import” is found, the syntax is expected to
match the import syntax. If it doesn’t. a parse error is thrown. If
import was not found, then it is instead a failure to parse an import:
there is no import.